fix: diff assertion truth, translate at the applier boundary, extract merge identity
Four review findings plus the structural residuals closed in-PR.

Same-id truth replacement is visible to the diff (P1). The identity diff
compared assertion presence by id, so a branch that hard-deletes an
endpoint (physically removing the row), recreates it, and imports the
same id for different truth diffed as EMPTY — the merge succeeded with
asserted: 0 while the target silently kept the truth the branch had
replaced. Shared ids are now compared by complete truth and a
replacement stages as retracted + new; because the applier never reuses
an ended row's id, the merge refuses typed instead of silently keeping
either side. Pinned deterministically (both matrix backends) and as a
property-law example.

Translation moved to the identity-applier boundary (P2). The typed
backstop now wraps exactly the applyIdentityMergeAtTarget call inside
applyMergePlan, and its classifier reads identity codes wherever the
applier puts them — top-level, details.code, or the transfer validator's
per-assertion details.issues[].code — and treats NodeNotFoundError as an
identity refusal, which is sound only at this boundary where a missing
node can mean nothing but a vanished assertion endpoint.

Law suite hardened (P2 x2). New BRANCH-EFFECT law: every truth a branch
holds at merge time must be applied with equal complete truth,
enumerated as dropped, retracted by some branch, or invalidated by an
endpoint deletion — silent loss of a branch belief now fails the suite
(the P1 repro satisfied every previous law). The operation alphabet
gains hard-delete/recreate and, in a new fold-profile lane, same-id
Robot peers joining and leaving Agent fold classes; both profiles run
the full law set on both backends. The generator's catch is now a strict
allowlist (IdentityContradictionError, NodeNotFoundError): any other
error fails the run instead of silently emptying the histories and
leaving the laws vacuously green.

Refactor: identity semantics of merge extracted to
graph-merge/merge-identity.ts (~1,150 lines out of merge.ts) — plan-time
derivation, contradiction simulation, and commit guards in one module
with a one-directional dependency from the orchestrator: functions take
a structural IdentityPlanSlice and the named IdentityPeerProbe type, so
the module never imports the plan type. describeCause moved to
graph-merge/errors.

Bonus fix the new report-coherence law surfaced: exact-duplicate staging
(two branches importing one identical row) reported the id as dropped
while applying it; dedupe now recognizes the identical row as ONE
assertion.

Docs: "Independent targets and assertion IDs" section in the identity
guide specifies the cross-lineage contract (one id one truth, truth-
matched retractions, replacement-as-conflict, in-transaction id
re-verification).

Validation: pnpm fix/typecheck/knip clean, API reports regenerated,
SQLite suite 6026 passed, Postgres suite 2733 passed, exit-gated. The
truth-replacement tests and the pinned law example fail against the
pre-fix diff (verified by revert).
